+++ src/taskranger/data/Task.hpp
@@ -44,6 +44,15 @@ public:
     const nlohmann::json getTaskJson() const;
     bool hasPublicIds();
     std::string getUUID();
+    template <typename T>
+    T getOrElse(const std::string& key, const T& def) {
+        auto& json = this->getTaskJson();
+        auto itr = json.find(key);
+        if (itr == json.end()) {
+            return def;
+        }
+        return *itr->get<T>();
+    }
 };
 
 } // namespace taskranger