+++ src/taskranger/data/attributes/AttribTypes.hpp
@@ -68,16 +68,18 @@ public:
 };
 
 class DateAttribute : public Attribute {
+private:
+    bool includeRelative = true;
+
 public:
     DateAttribute() {
         this->type = FieldType::DATE;
     }
 
-    virtual void validate(nlohmann::json& /* attribValue */) override {
-        // TODO + conversion (?)
-    }
+    virtual void validate(nlohmann::json& /* attribValue */) override {}
 
     virtual void modify(nlohmann::json& task, const std::string& key, const std::string& input) override;
+    virtual std::variant<std::string, tabulate::Table> print(const Task& task) override;
 };
 
 } // namespace taskranger