+++ src/taskranger/util/TableUtil.cpp
@@ -1,17 +1,59 @@
 #include "TableUtil.hpp"
 #include "tabulate/color.hpp"
+#include "taskranger/data/Environment.hpp"
 #include "taskranger/util/StrUtil.hpp"
 #include <algorithm>
 
 namespace taskranger {
 
-tabulate::Table TableUtil::renderTasks(
-        std::vector<std::shared_ptr<Task>>& renderTarget, std::map<std::string, int> keyPriority) {
-    // TODO: Stress-test this in actual cases and cap the size
+namespace TableUtil {
+
+TableBuilder::TableBuilder() {}
+
+TableBuilder& TableBuilder::withKeys(const std::vector<std::string>& keys) {
+    this->keys = keys;
+    for (auto& key : keys) {
+        columns[key] = {};
+    }
+    return *this;
+}
+
+tabulate::Table TableBuilder::build(std::vector<std::shared_ptr<Task>>& tasks) {
+    std::vector<TableRow> keysWithValues;
+    auto& env = *Environment::getInstance();
+    for (auto& taskPtr : tasks) {
+        auto& task = *taskPtr;
+        auto& json = task.getTaskJson();
+        for (auto& key : keys) {
+            auto itr = json.find(key);
+            if (itr == json.end()) {
+                columns[key].push_back("");
+                continue;
+            }
+            std::shared_ptr<Attribute> attribPtr = nullptr;
+            try {
+                attribPtr = env.getAttribute(key);
+
+            } catch (std::string&) {}
+            if (!attribPtr) {
+                columns[key].push_back(StrUtil::toString(*itr, " ", ""));
+            } else {
+                columns[key].push_back(attribPtr->print(task));
+            }
+            // This system is used to make sure only keys with at least one value are printed.
+            // The rest shouldn't be printed
+            if (std::find_if(keysWithValues.begin(), keysWithValues.end(), [&key = key](const auto& arg) {
+                    // The following exception should never be thrown.
+                    if (!std::holds_alternative<std::string>(arg))
+                        throw std::runtime_error("This is why we can't have nice things");
+                    return std::get<std::string>(arg) == key;
+                }) == keysWithValues.end()) {
+                keysWithValues.push_back(key);
+            }
+        }
+    }
     tabulate::Table table;
     // clang-format off
-    // Clear the edges for a slightly slicker feel.
-    // (Borders don't make sense for this specific table)
     table.format()
         .border_top("")
         .border_bottom("")
@@ -19,75 +61,50 @@ tabulate::Table TableUtil::renderTasks(
         .border_right(" ")
         .corner("");
     // clang-format on
+    std::vector<TableRow> tfKeys;
 
-    std::vector<TableRow> keys;
-
-    for (auto& task : renderTarget) {
-        for (auto& [k, v] : task->getTaskJson().items()) {
-            // Because std::variant doesn't define an operator==, this is a hack to work around
-            // that.
-            if (std::find_if(keys.begin(), keys.end(), [&key = k](const auto& arg) {
-                    // The following exception should never be thrown.
-                    if (!std::holds_alternative<std::string>(arg))
-                        throw std::runtime_error("This is why we can't have nice things");
-                    return std::get<std::string>(arg) == key;
-                }) == keys.end()) {
-                keys.push_back(k);
-            }
+    for (auto& fuckingBullshitFormat : keysWithValues) {
+        std::shared_ptr<Attribute> attribute = nullptr;
+        std::string key = std::get<std::string>(fuckingBullshitFormat);
+        try {
+            attribute = env.getAttribute(key);
+        } catch (std::string&) {}
+        if (!attribute) {
+            tfKeys.push_back(key);
+            continue;
         }
+        tfKeys.push_back(attribute->getLabel());
     }
+    table.add_row(tfKeys);
+    // Underline the keys to highlight them
+    table[0].format().font_style({tabulate::FontStyle::underline});
 
-    // Sorts the keys by priority. Unprioritized ones are sorted slightly randomly, but hopefully consistently.
-    // possibly stupid code
-    std::sort(keys.begin(), keys.end(), [&keyPriority = keyPriority](const auto& rawA, const auto& rawB) -> bool {
-        auto a = std::get<std::string>(rawA);
-        auto b = std::get<std::string>(rawB);
-
-        auto itA = keyPriority.find(a);
-        auto itB = keyPriority.find(b);
-        if (itA == keyPriority.end() && itA == itB)
-            return false;
-        int valA, valB;
-        if (itA != keyPriority.end())
-            valA = itA->second;
-        else
-            valA = 5;
-        if (itB != keyPriority.end())
-            valB = itB->second;
-        else
-            valB = 5;
-        return valA < valB;
-    });
-
-    table.add_row(keys);
-    // Modify the description to stay narrow.
-    // This might need to be tweaked based on terminal width.
-    // See TermUtils.hpp for a function that gets the cols (width)
-    // of the terminal.
+    // Limit the description size (temporary)
     for (auto& cell : table.row(0)) {
-        if (cell.get_text() == "description") {
+        if (cell.get_text() == "Description") {
             cell.format().width(60);
             break;
         }
     }
 
-    table[0].format().font_style({tabulate::FontStyle::underline});
+    std::vector<std::vector<TableRow>> rows;
 
-    for (auto& taskObj : renderTarget) {
-        auto& task = taskObj->getTaskJson();
-        std::vector<TableRow> row;
-        for (auto& key : keys) {
-            auto value = task.find(std::get<std::string>(key));
-            if (value == task.end()) {
-                row.push_back("");
-            } else {
-                row.push_back(StrUtil::toString(*value));
+    // Iterate keysWithValues to preserve order
+    for (auto& key : keysWithValues) {
+        auto& values = columns.at(std::get<std::string>(key));
+
+        for (size_t i = 0; i < values.size(); i++) {
+            if (i == rows.size()) {
+                rows.push_back({});
             }
+            rows.at(i).push_back(values.at(i));
         }
+    }
+    for (auto& row : rows) {
         table.add_row(row);
     }
-    // Minor hack; the index is required, but the table class doesn't expose
-    // any sizes, or have index-based access.
+
+    // Pretty background colors
     size_t idx = 0;
     for (auto& row : table) {
         // grey is black according to the source code.
@@ -98,5 +115,7 @@ tabulate::Table TableUtil::renderTasks(
     }
     return table;
 }
+
+} // namespace TableUtil
 
 } // namespace taskranger